% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.ww3.R
\name{download.ww3}
\alias{download.ww3}
\title{download ww3 function}
\usage{
download.ww3(var = "hs", anos, meses, dir = getwd())
}
\description{
This function allows you to download ww3 data.
}
\examples{
download.ww3()
}
\keyword{ww3}
