% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.values.R
\name{extract.values}
\alias{extract.values}
\title{extract values from ww3 data function}
\usage{
extract.values(dir, x, y, ncluster = 1)
}
\description{
This function allows you to extract values from ww3 data.
}
\examples{
extract.values()
}
\keyword{ww3}
